+++ src-tauri/src/config.rs
@@ -1,10 +1,6 @@
-use std::sync::{
-    atomic::{AtomicBool, AtomicU8, Ordering},
-    Arc,
-};
-
 use anyhow::Result;
 use once_cell::sync::Lazy;
+use parking_lot::Mutex;
 use serde::{Deserialize, Serialize};
 
 use crate::util;
@@ -17,71 +13,114 @@ pub struct Config {
     /// 模式, true: 镜像模式, false: 直连模式
     ///
     /// Mode, true: mirror mode, false: direct mode
+    #[serde(default = "mode_default")]
     pub mode: bool,
 
     /// 快捷键
     ///
     /// Key, 0: shift, 1: ctrl, 2: caps
+    #[serde(default = "key_default")]
     pub key: u8,
+
+    /// 主题
+    ///
+    /// Theme
+    #[serde(default = "theme_default")]
+    pub theme: String,
 }
 
-/// API模式
-///
-/// API mode
-pub static MODE: Lazy<Arc<AtomicBool>> = Lazy::new(|| Arc::new(AtomicBool::new(true)));
+fn mode_default() -> bool {
+    true
+}
 
-/// 快捷键配置
-///
-/// Key config
-pub static KEY: Lazy<Arc<AtomicU8>> = Lazy::new(|| Arc::new(AtomicU8::new(0)));
+fn key_default() -> u8 {
+    0
+}
+
+fn theme_default() -> String {
+    "dark".to_string()
+}
+
+pub static CONFIG: Lazy<Mutex<Config>> = Lazy::new(|| Mutex::new(load()));
 
 /// 读取配置
 ///
 /// read config
-pub fn load() {
+pub fn load() -> Config {
     let exe_dir = util::get_exe_dir();
     let path = exe_dir.join("config.json");
     // 检测文件是否存在
     // Check if file exists
     let exists = path.try_exists();
-    if exists.is_ok() && exists.expect("Failed to check if file exists") {
-        let config = std::fs::read_to_string(path).expect("Failed to read config");
-        let config = serde_json::from_str::<Config>(&config).expect("Failed to parse config");
-        MODE.store(config.mode, Ordering::SeqCst);
-        KEY.store(config.key, Ordering::SeqCst);
+    let json = if exists.is_ok() && exists.expect("Failed to check if file exists") {
+        std::fs::read_to_string(path).expect("Failed to read config")
     } else {
-        MODE.store(true, Ordering::SeqCst);
-        KEY.store(0, Ordering::SeqCst);
-    }
+        "{}".to_string()
+    };
+
+    serde_json::from_str::<Config>(&json).expect("Failed to parse config")
 }
 
 /// 保存配置
 ///
 /// save config
 fn save() -> Result<()> {
-    let config = Config {
-        mode: MODE.load(Ordering::SeqCst),
-        key: KEY.load(Ordering::SeqCst),
-    };
     let exe_dir = util::get_exe_dir();
     let path = exe_dir.join("config.json");
-    let config = serde_json::to_string_pretty(&config).expect("Failed to serialize config");
+    let config =
+        serde_json::to_string_pretty(&CONFIG.lock().clone()).expect("Failed to serialize config");
     std::fs::write(path, config).expect("Failed to write config");
     Ok(())
+}
+
+/// 获取全局配置
+///
+/// get global config
+pub fn config() -> Config {
+    CONFIG.lock().clone()
+}
+
+/// 获取模式
+///
+/// get mode
+pub fn mode() -> bool {
+    CONFIG.lock().mode
+}
+
+/// 获取快捷键
+///
+/// get key
+pub fn key() -> u8 {
+    CONFIG.lock().key
+}
+
+/// 获取主题
+///
+/// get theme
+pub fn theme() -> String {
+    CONFIG.lock().theme.clone()
 }
 
 /// 切换模式
 ///
 /// switch mode
-pub fn mode(mode: bool) {
-    MODE.store(mode, Ordering::SeqCst);
+pub fn set_mode(mode: bool) {
+    CONFIG.lock().mode = mode;
     save().expect("Failed to save config after switch mode");
 }
 
 /// 切换快捷键
 ///
 /// switch key
-pub fn key(key: u8) {
-    KEY.store(key, Ordering::SeqCst);
+pub fn set_key(key: u8) {
+    CONFIG.lock().key = key;
     save().expect("Failed to save config after switch key");
+}
+
+/// 切换主题
+///
+/// switch theme
+pub fn set_theme(theme: &str) {
+    CONFIG.lock().theme = theme.to_string();
+    save().expect("Failed to save config after switch theme");
 }